Renamed as_variant_descs to variant_descs
