+++ sylvia-derive/src/lib.rs
@@ -6,7 +6,6 @@ use quote::quote;
 use syn::fold::Fold;
 use syn::{parse2, parse_quote, ItemImpl, ItemTrait, Path};
 
-mod as_variant_descs;
 pub(crate) mod check_generics;
 mod input;
 mod message;
@@ -16,6 +15,7 @@ mod remote;
 mod strip_generics;
 mod strip_input;
 mod utils;
+mod variant_descs;
 
 use strip_input::StripInput;
